+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.vashisthg.androidl" >
+    package="com.ophio.androidl" >
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,7 @@
         android:theme="@style/Theme.MyApp"
         android:label="@string/app_name" >
         <activity
-            android:name=".view.MainActivity"
+            android:name="com.ophio.androidl.view.MainActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -16,9 +16,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.ActivityTransitionsDemoActivity"/>
-        <activity android:name=".view.TransitionsActivity"/>
-        <activity android:name=".view.RevealDemoActivity"/>
+        <activity android:name="com.ophio.androidl.view.ActivityTransitionsDemoActivity"/>
+        <activity android:name="com.ophio.androidl.view.TransitionsActivity"/>
+        <activity android:name="com.ophio.androidl.view.RevealDemoActivity"/>
 
     </application>
