+++ app/src/main/java/com/ophio/androidl/PalleteTransformation.java
@@ -1,4 +1,4 @@
-package com.vashisthg.androidl;
+package com.ophio.androidl;
 
 import android.graphics.Bitmap;
 import android.support.v7.graphics.Palette;
