+++ app/src/main/java/com/ophio/androidl/view/DemoEntity.java
@@ -1,4 +1,4 @@
-package com.vashisthg.androidl.view;
+package com.ophio.androidl.view;
 
 import android.support.annotation.DrawableRes;
