+++ app/src/main/res/layout/fragment_demo.xml
@@ -3,7 +3,7 @@
              xmlns:app="http://schemas.android.com/apk/res-auto"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.vashisthg.androidl.view.fragment.DemoFragment">
+             tools:context="com.ophio.androidl.view.fragment.DemoFragment">
 
 
     <org.lucasr.twowayview.widget.TwoWayView
