+++ app/src/main/res/layout/navdrawer.xml
@@ -14,7 +14,7 @@
    limitations under the License.
 -->
 
-<com.vashisthg.androidl.widget.ScrimInsetsScrollView
+<com.ophio.androidl.widget.ScrimInsetsScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="@dimen/navdrawer_width"
@@ -35,4 +35,4 @@
     </LinearLayout>
 
 
-</com.vashisthg.androidl.widget.ScrimInsetsScrollView>
+</com.ophio.androidl.widget.ScrimInsetsScrollView>
